Update in-case-of-issues.md
If user does not run this command as www-data, it will result in suggestions following file permission warnings which will leave the folders with user:user permissions instead of user:www-data. Although healthcheck will pass, errors will be created because (for example, because tmp folder is not writable by www-data).